Generic multi_panel option for FormatCBFMiniPilatus
